+++ app/controllers/design_options_controller.rb
@@ -182,7 +182,7 @@ class DesignOptionsController < ApplicationController
       :show_current_button, :show_seconds, :hard_minimum, :hard_maximum, :soft_minimum, :soft_maximum, :domain_id,
       :date_hard_maximum, :date_hard_minimum, :date_soft_maximum, :date_soft_minimum, :date_format, :time_of_day_format,
       :time_duration_format, :calculation, :calculated_format, :hide_calculation, :units, :multiple_rows,
-      :default_row_number, :alignment,
+      :default_row_number, :alignment, :disallow_future_dates,
       { grid_tokens: [:variable_id] }, { questions: [:question_name, :question_type] }, :autocomplete_values
     )
   end