+++ app/controllers/variables_controller.rb
@@ -131,7 +131,7 @@ class VariablesController < ApplicationController
       :hard_minimum, :hard_maximum, :soft_minimum, :soft_maximum,
       # For Dates
       :date_hard_maximum, :date_hard_minimum, :date_soft_maximum,
-      :date_soft_minimum, :date_format,
+      :date_soft_minimum, :date_format, :disallow_future_dates,
       # For Date, Time of Day
       :show_current_button,
       # For Time of Day