+++ db/migrate/20190811192356_add_disallow_future_dates_to_variable.rb
@@ -1,0 +1,5 @@
+class AddDisallowFutureDatesToVariable < ActiveRecord::Migration[6.0]
+  def change
+    add_column :variables, :disallow_future_dates, :boolean, null: false, default: false
+  end
+end